chore: We have split the types and the interfaces in different files in order to make more readable the code
